+++ lib/wa_tools.dart
@@ -9,3 +9,5 @@ export 'src/tools/convertor/string_validator.dart';
 export 'src/tools/convertor/convert_size.dart';
 export 'src/tools/convertor/query_string.dart';
 export 'src/tools/convertor/map_tools.dart';
+export 'src/tools/extensions/directory.dart';
+export 'src/tools/extensions/file.dart';